src/api/requirements.txt
fastapi
fastapi[standard]
requests
gunicorn==21.2.0
azure-cosmos
azure-search-documents==11.4.0
azure-keyvault-secrets
azure-monitor-opentelemetry-exporter
aiohttp==3.9.5
python-dotenv==1.0.1
azure-cognitiveservices-speech==1.37.0
prompty[azure]==0.1.24
jsonlines
nbconvert
pandas
azure-identity==1.17.1
jupyter
Jinja2==3.1.4
azure-ai-inference[opentelemetry]
opentelemetry-sdk 
azure-monitor-opentelemetry
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation
opentelemetry-exporter-otlp-proto-grpc
azure-ai-evaluation
azure-monitor-query